@@ -5,9 +5,9 @@ using Facteur.TemplateProviders.IO;
 
 namespace Facteur.Examples
 {
-    class Program
+    internal class Program
     {
-        static async Task Main(string[] args)
+        private static async Task Main(string[] args)
         {
             Console.WriteLine("Sending welcome mail using a GMAIL account");
 
@@ -25,30 +25,21 @@ namespace Facteur.Examples
 
         private static async Task SendConfirmationMail(string from, string fromPassword, string customerMail, string customerName)
         {
-            // Construct the e-mail model
-            EmailComposer<WelcomeMailModel> composer = new EmailComposer<WelcomeMailModel>();
-            EmailRequest<WelcomeMailModel> request = composer
-                .SetModel(new WelcomeMailModel { Email = customerMail, Name = customerName })
+            // When using a gmail account: https://stackoverflow.com/a/25215834/1842261
+            SmtpCredentials credentials = new("smtp.gmail.com", "587", "false", "true", from, fromPassword);
+
+            // The composition:
+            // - Compiler: Scriban template syntax,
+            // - Provider files in this project's directory
+            // - Template finder: finding templates in the app directory via the name of the view models
+            EmailComposer composer = new(new ScribanCompiler(), new AppDirectoryTemplateProvider("Templates", ".sbnhtml"), new ViewModelTemplateResolver());
+
+            IMailer mailer = new SmtpMailer(credentials, composer);
+            await mailer.SendMailAsync(x => x
                 .SetSubject("Welcome to the company!")
                 .SetFrom("info@facteur.com")
                 .SetTo(customerMail)
-                .Build();
-
-            // Creating the body using the following components:
-            // * Use Scriban as the template syntax
-            // * Store templates in the project under the 'Templates' application directory
-            // * Map e-mail templates by the mail model name (WelcomeMailModel -> Welcome.sbnhtml)
-            IMailBodyBuilder builder = new MailBodyBuilder(
-                new ScribanCompiler(),
-                new AppDirectoryTemplateProvider("Templates", ".sbnhtml"),
-                new ViewModelTemplateResolver());
-
-            EmailRequest populatedRequest = await builder.BuildAsync(request);
-
-            // When using a gmail account: https://stackoverflow.com/a/25215834/1842261
-            SmtpCredentials credentials = new("smtp.gmail.com", "587", "false", "true", from, fromPassword);
-            IMailer mailer = new SmtpMailer(credentials);
-            await mailer.SendMailAsync(populatedRequest);
+                .BuildAsync(new WelcomeMailModel { Email = customerMail, Name = customerName }));
         }
     }
 }
